Using entities for the user is weird, so let's switch back to a more traditional reducer pattern.
