+++ src/shared/Entities/schema.js
@@ -3,7 +3,6 @@ import { schema } from 'normalizr';
 
 // User
 export const user = new schema.Entity('users');
-export const loggedInUser = user;
 
 // Uploads
 export const upload = new schema.Entity('uploads');
