--- src/shared/User/ducks.js
@@ -1,86 +1,0 @@
-import * as Cookies from 'js-cookie';
-import * as helpers from 'shared/ReduxHelpers';
-import { isMilmoveSite, isOfficeSite, isTspSite } from 'shared/constants';
-import { GetLoggedInUser } from './api.js';
-import { normalize } from 'normalizr';
-import { pick } from 'lodash';
-
-import { ordersArray } from 'shared/Entities/schema';
-import { addEntities } from 'shared/Entities/actions';
-import { getShipment } from 'shared/Entities/modules/shipments';
-
-const getLoggedInUserType = 'GET_LOGGED_IN_USER';
-
-export const GET_LOGGED_IN_USER = helpers.generateAsyncActionTypes(getLoggedInUserType);
-
-const getLoggedInActions = helpers.generateAsyncActions(getLoggedInUserType);
-export const loadLoggedInUser = () => {
-  return function(dispatch) {
-    const userInfo = getUserInfo();
-    if (!userInfo.isLoggedIn) return Promise.resolve();
-    dispatch(getLoggedInActions.start());
-    return GetLoggedInUser()
-      .then(response => {
-        if (response.service_member) {
-          const data = normalize(response.service_member.orders, ordersArray);
-          if (data.entities.shipments) {
-            const shipmentIds = Object.keys(data.entities.shipments);
-            shipmentIds.map(id => dispatch(getShipment(id)));
-          }
-
-          // Only store addresses in a normalized way. This prevents
-          // data duplication while we're using both Redux approaches.
-          const filtered = pick(data.entities, ['addresses']);
-          dispatch(addEntities(filtered));
-        }
-        return dispatch(getLoggedInActions.success(response));
-      })
-      .catch(error => dispatch(getLoggedInActions.error(error)));
-  };
-};
-
-// the results of the api call will be handled by other reducers. This just lets us know app has loaded initial data
-export const loggedInUserReducer = (state = {}, action) => {
-  switch (action.type) {
-    case GET_LOGGED_IN_USER.start:
-      return {
-        ...state,
-        hasSucceeded: false,
-        hasErrored: false,
-        isLoading: true,
-      };
-    case GET_LOGGED_IN_USER.success:
-      return {
-        ...state,
-        isLoading: false,
-        hasErrored: false,
-        hasSucceeded: true,
-      };
-    case GET_LOGGED_IN_USER.error:
-      return {
-        ...state,
-        isLoading: false,
-        hasErrored: true,
-        hasSucceeded: false,
-        error: action.error,
-      };
-    default:
-      return state;
-  }
-};
-
-function getUserInfo() {
-  // The prefix should match the lowercased application name set in the server session
-  let cookiePrefix = (isMilmoveSite && 'mil') || (isOfficeSite && 'office') || (isTspSite && 'tsp') || '';
-  const cookieName = cookiePrefix + '_session_token';
-  const cookie = Cookies.get(cookieName);
-  return {
-    isLoggedIn: !!cookie,
-  };
-}
-
-const userReducer = (state = getUserInfo(), action) => {
-  return getUserInfo();
-};
-
-export default userReducer;
